refactor(gateway): pull shared ValidateInternalAsync up into VirtualKeyServiceBase
Direct and Cached key services carried ~60-line identical validation flows,
differing only in the hash lookup (repository vs cache). The cached service
also picks up the success-path LogDebug it had drifted away from.
