+++ Services/ConduitLLM.Gateway/Services/DirectApiVirtualKeyService.cs
@@ -56,67 +56,17 @@ namespace ConduitLLM.Gateway.Services
             return ValidateInternalAsync(key, requestedModel, checkBalance: true);
         }
 
-        private async Task<VirtualKeyValidationOutcome> ValidateInternalAsync(
+        private Task<VirtualKeyValidationOutcome> ValidateInternalAsync(
             string key,
             string? requestedModel,
             bool checkBalance)
         {
-            if (string.IsNullOrWhiteSpace(key))
-            {
-                Logger.LogWarning("Empty key provided for virtual key validation");
-                return VirtualKeyValidationOutcome.Failure(
-                    VirtualKeyValidationFailureCodes.MissingKey,
-                    401,
-                    "Virtual key is required.");
-            }
-
-            try
-            {
-                var keyHash = VirtualKeyUtilities.HashKey(key);
-                Logger.LogDebug("Validating key ({ValidationMode}): {KeyPrefix}, Hash: {Hash}",
-                    checkBalance ? "balance" : "authentication",
-                    LoggingSanitizer.S(ConduitLLM.Core.Utilities.SpanHelper.MaskSecret(key)),
-                    keyHash);
-
-                var virtualKey = await VirtualKeyRepository.GetByKeyHashAsync(keyHash);
-                if (virtualKey == null)
-                {
-                    Logger.LogWarning("No matching virtual key found for hash: {Hash}", keyHash);
-                    return VirtualKeyValidationOutcome.Failure(
-                        VirtualKeyValidationFailureCodes.KeyNotFound,
-                        401,
-                        "Virtual key was not found.");
-                }
-
-                var result = await VirtualKeyValidationHelper.ValidateVirtualKeyAsync(
-                    virtualKey,
-                    requestedModel,
-                    checkBalance,
-                    checkBalance ? GroupRepository : null,
-                    Logger,
-                    checkBalance ? _batchSpendService : null);
-
-                if (!result.IsValid)
-                {
-                    Logger.LogWarning("Virtual key {KeyId} validation failed: {Reason}",
-                        virtualKey.Id, result.Reason ?? "unknown");
-                }
-                else
-                {
-                    Logger.LogDebug("Virtual key {KeyId} validated successfully for model: {Model}",
-                        virtualKey.Id, LoggingSanitizer.S(requestedModel ?? "any"));
-                }
-
-                return result;
-            }
-            catch (Exception ex)
-            {
-                Logger.LogError(ex, "Error validating virtual key");
-                return VirtualKeyValidationOutcome.Failure(
-                    VirtualKeyValidationFailureCodes.ValidationError,
-                    500,
-                    "Virtual key validation failed.");
-            }
+            return ValidateVirtualKeyInternalAsync(
+                key,
+                requestedModel,
+                checkBalance,
+                hash => VirtualKeyRepository.GetByKeyHashAsync(hash),
+                _batchSpendService);
         }
 
         /// <inheritdoc />
